+++ eng/pipelines/cadl-ci.yml
@@ -63,7 +63,7 @@ stages:
           displayName: 'Install Dependencies for CADL csharp'
           inputs:
             command: install
-            workingDir: $(Build.SourcesDirectory)/src/CADL.Extension/Emitter.Csharp
+            workingDir: $(Build.SourcesDirectory)
         - task: Npm@1
           displayName: 'Build CADL csharp emitter'
           inputs:
