+++ test/AutoRest.TestServer.Tests/Infrastructure/TestServerLowLevelTestBase.cs
@@ -2,14 +2,10 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
 using System.Threading.Tasks;
 using Azure;
 using Azure.Core;
 using Azure.Core.Pipeline;
-using Microsoft.AspNetCore.Http;
 using NUnit.Framework;
 
 namespace AutoRest.TestServer.Tests.Infrastructure
@@ -42,39 +38,6 @@ namespace AutoRest.TestServer.Tests.Infrastructure
         public Task Test(Func<Uri, Task> test, bool ignoreScenario = false)
         {
             return Test(GetScenarioName(), test, ignoreScenario);
-        }
-
-        public async Task Test(Func<Uri, Task> test, TestServerType[] types)
-        {
-            if (types.Contains(TestServerType.CadlRanch))
-            {
-                var server = CadlRanchServerSession.Start();
-
-                try
-                {
-                    await test(server.Host);
-                }
-                catch (Exception ex)
-                {
-                    try
-                    {
-                        await server.DisposeAsync();
-                    }
-                    catch (Exception disposeException)
-                    {
-                        throw new AggregateException(ex, disposeException);
-                    }
-
-                    throw;
-                }
-
-                await server.DisposeAsync();
-            }
-
-            if (types.Contains(TestServerType.TestServer))
-            {
-                await Test(test);
-            }
         }
 
         private async Task Test(string scenario, Func<Uri, Task> test, bool ignoreScenario = false)
