+++ test/CadlRanchProjects.Tests/api-key.cs
@@ -10,20 +10,20 @@ using System.Threading.Tasks;
 
 namespace CadlRanchProjects.Tests
 {
-    public class api_key : TestServerLowLevelTestBase
+    public class api_key : CadlRanchTestBase
     {
         [Test]
         public Task Authentication_ApiKey_valid() => Test(async (host) =>
         {
             Response response = await new ApiKeyClient(new AzureKeyCredential("valid-key"), host, null).ValidAsync();
             Assert.AreEqual(204, response.Status);
-        }, new[] { TestServerType.CadlRanch });
+        });
 
         [Test]
         public Task Authentication_ApiKey_invalid() => Test(async (host) =>
         {
             Response response = await new ApiKeyClient(new AzureKeyCredential("valid-key"), host, null).InvalidAsync();
             Assert.AreEqual("invalid-api-key", InvalidAuth.FromResponse(response).Error);
-        }, new[] { TestServerType.CadlRanch });
+        });
     }
 }
